refactor tagMap type folder location
